Draw ROI boxes
Draw boxes around the ROIs detectd for a camera event
