+++ src/services/image.js
@@ -1,0 +1,8 @@
+const IMAGE_W = 1024
+const IMAGE_H = 768
+
+const SCALE = 0.5
+
+const scaledPt = (pt) => pt * SCALE
+
+export {IMAGE_W, IMAGE_H, SCALE, scaledPt}
